Skip node expansion of ops that cause issues for mnist_train_test.
The following ops cause the mnist_train_test to fail on TPU with the MLIR
bridge when node expansion is done on them:

NoOp - hits a deadlock condition in shape inference.
IdentityOp - leads to wrong results in the test.
OptionalHasValue - not legalizable TF to XLA.
OptionalGetValue - not legalizable TF to XLA.

As a temporary workaround, skip node expansion of these ops. With this
workaround, the mnist_train_test now passes on TPU with the MLIR bridge.

PiperOrigin-RevId: 347680741
Change-Id: I39702615b9e12d6dfbdc7b3e4689542bbc16f5a9