Update TF BroadcastGradientArgs op folder to match TF behavior
In case both the operands are equal, TensorFlow choses to not reduce any of the dimension and outputs zero sized vector. The folder outputs dimensions that has ones in both the operands. For example, [1, 2] and [1, 2] will compute [0] and [0] in the folder but TensorFlow outputs [] and [].

This doesn't actually matter at the usage as reducing dimensions of size one has no effect but maintaining the TF behavior to avoid shape verification failures.

PiperOrigin-RevId: 347690274
Change-Id: I6851116cc0f3d60a362494e699b8d12fc79a1979