Update the external license check to be close to the internal check.
Manually updated the licenses in the files that are not excluded from a
license check in the internal CI system, and removed these exclusions in
the external CI system as well.

The external exclusions that are not excluded internally are:
 * .bzl
 * .h5
 * .tpl
 * .properties

Will will fix these remaining discrepancies incrementally, as and when a
PR touches such a file and runs afoul of the internal checks.

See http://b/175315163#comment7 for more details.