Update the style checks to also include the reference kernels.
Prior to this change we were only looking at code within the
tensorflow/lite/micro directory. However, TFLM does share the reference
implementations with Lite and so we are also checking for license and
formatting in those files.

This should help detect errors faster, for example, PR
https://github.com/tensorflow/tensorflow/pull/45814 missed adding a
license to the header but that wasn't detected until the PR was improted
internally.

See http://b/169948621 and http://b/175315163 for more details.