Add a new test for overriding an existing kernel
This path also fixed one issue that a tf attribute has type information, and
the type can be different from Python. For example, if a scalar is used to
construct a tf constant tensor, the scalar will be mapped to mlir scalar
attribute, but needs to be converted to a tensor attribute when tf constant
operation is constructed. Then the attribute needs to be converted explicitly
when the op is raised to TF.

PiperOrigin-RevId: 350221650
Change-Id: I3a9be6805a261b720b3a9b8e63131db4be6a6a11