Add cluster ops-by-policy pass
This pass clusters ops according to the policy specified by the pass options. The only currently supported option is 'oplist'. This option allows to specify the names of the ops that should be clustered when they form a simple def-use chain, that is, the next op is the only user of the previous op result. The pass identifies sequences of such ops and extracts them into clusters, represented by tf_device::ClusterOp.

This pass will be used to support flexibly policy in defining operations to be compiled in a single kernel for CPU code generation. The 'oplist' policy will be used to support user-guided fusion, where a specified sequence of operations is replaced by a single fused jitted kernel.

PiperOrigin-RevId: 350715123
Change-Id: I511c38ab9a9f2f40dd8df4f67d4747e25e668b26