Do not cache variable scopes in legacy_tf_layers.base.Layer when Eager is enabled (regardless of the current context).
Caching the variable scope causes the layer to be "poisoned" when used within a tf.function, since if the layer is called for the first time inside a tf.function, then a FuncGraph scope is captured and then re-entered on every subsequent call. This caching was simply a graph-building (Python) performance optimization and can be skipped if Eager is enabled.

PiperOrigin-RevId: 351286068
Change-Id: Ia82958b8fca83bac36f6bc3ce4dd08a8e5011ca0