Do not hang in GPU tf.math.pow if RHS is a negative int.
Before, if the RHS to tf.math.pow in the GPU kernel was negative and the inputs were int64, the GPU kernel would hang. Eigen expects the RHS to be nonnegative and will go into an infinite loop otherwise. The CPU kernel already checks this and raises an error if the RHS is negative, but we cannot easily raise errors in GPU kernels.

PiperOrigin-RevId: 351461465
Change-Id: Ia474fa08b7e4d830c3d4a2f04bd1cb6cda83736d