Enable clang+bazel as part of the TFLM github CI.
We patch some of the TF workspace and BUILD files to reduce the number
of downloads as part of a bazel build.

We can revisit this decision if the maintenance overhead from
this patching becomes too high.

Also, there may be an opportunity to further reduce the downloads, but
this change has made a decision to leave the TfLite build files
unchanged and incur the cost of the additional downloads that are needed
for TfLite but not for TfLite Micro.

Fixes #46465
Fixes http://b/177672856